Complete BLE Functionality App Integration
This final commit is all the features required for the app, without
stable, automatic connection and disconnection to a remembered device,
unless it's manually forgotten. One small testing feature left inside at
line 442: <||name=="No Name">. This, as well as the debug box should be
removed for the final implementation.