Prevented App Crashing from accessing null Queue
The main.cs class was not properly referencing the data in the
controller.cs (bluetooth) class, thus crashing the app when a request
for this data was instantiated. This has since been fixed. Next work is
to solve the issue I informed Leah of, and implement checks to prevent
the app entering a state which will crash it if there is no BLE device
connected.